+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 ### Added
 
 * `Faker.Address.Es` and `Faker.Address.En` [@vbrazo][]
+* Add doctests for Superhero [@vbrazo][]
+* Add doctests for Team [[@vbrazo][]]
+* `Faker.Name.PtBr` and `Faker.Name.Es` [@vbrazo][]
 * Add doctests for Commerce [[@vbrazo][]]
 * Add doctests for StarWars [[@vbrazo][]]
 * Add doctests for Company [[@vbrazo][]]